feat: extract type
